fixed dependencies, logging, changed some datatypes in weatherwise

+++ project/src/java/weatherwise/ForecastWeatherReport.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ForecastWeatherReport {
-    private Double temperature;
-    private Integer humidity;
-    private Integer pressure;
+    private Long temperature;
+    private Long humidity;
+    private Long pressure;
 }
